- Added a warning to the uninstaller (so you won't remove any precious   files!)

+++ Uninstaller.sh
@@ -2,8 +2,13 @@
 
 WHITELISTED_FILES=("LICENSE" "README.md" "Installer.sh" "Uninstaller.sh")
 
-for file in *; do 
-  if [[ ! "${WHITELISTED_FILES[@]}" =~ "$file" ]]; then
-    rm -rf "$file"
-  fi 
-done
+printf "WARNING! THIS SCRIPT WILL REMOVE ALL FILES FROM THIS DIRECTORY ($(pwd)).\nDO YOU WANT TO PROCEED? [y/n]: "
+read warn 
+
+if [[ $(echo "$warn" | awk '{print tolower($0)}') = "y" ]]; then
+  for file in *; do 
+    if [[ ! "${WHITELISTED_FILES[@]}" =~ "$file" ]]; then
+      rm -rf "$file"
+    fi 
+  done
+fi
